Added authorization headers to all api calls

+++ src/hooks/queries/useEntryDeletion.tsx
@@ -3,29 +3,37 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { endPointURL, getQueryKey } from "@/utils/constant-vars";
 import axios from "axios";
 import { ServerResponseObj } from "@/utils/types";
+import getUserLogin from "@/utils/user-cookies/get-user-login";
 
 export default function useEntryDeletion() {
-    const mainQueryClient = useQueryClient();
+  const mainQueryClient = useQueryClient();
 
-    const {mutateAsync, isSuccess, error} = useMutation({
-        mutationFn: deleteFinancialEntry,
-        onSuccess: invalidateQueries
-    })
+  const { mutateAsync, isSuccess, error } = useMutation({
+    mutationFn: deleteFinancialEntry,
+    onSuccess: invalidateQueries,
+  });
 
-    async function deleteFinancialEntry(entryID: number) {
-        const response = await axios({
-            method: "delete",
-            url: `${endPointURL}/entries/remove-financial-entry`,
-            data: {
-                entryID
-            }
-        })
-        return response.data;
-    }
+  async function deleteFinancialEntry(entryID: number) {
+    const userSession = await getUserLogin();
+    if (userSession === null || userSession === undefined) return null;
+    const headers = {
+      authorization: userSession,
+    };
 
-    async function invalidateQueries() {
-        mainQueryClient.invalidateQueries({queryKey: [getQueryKey]});
-    }
+    const response = await axios({
+      method: "delete",
+      headers,
+      url: `${endPointURL}/entries/remove-financial-entry`,
+      data: {
+        entryID,
+      },
+    });
+    return response.data;
+  }
 
-    return {mutateAsync, isSuccess, error};
+  async function invalidateQueries() {
+    mainQueryClient.invalidateQueries({ queryKey: [getQueryKey] });
+  }
+
+  return { mutateAsync, isSuccess, error };
 }
